be/pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = "kac be system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "mysql-connector-python>=9.5.0",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.38.0",
]
